# Replaced deprecated x/tools/go/loader with x/tools/go/packages