--- a/.github/workflows/go-test.yml
+++ b/.github/workflows/go-test.yml
@@ -9,7 +9,7 @@ jobs:
   test:
     strategy:
       matrix:
-        go-version: [1.15]
+        go-version: [1.16]
         platform: [ubuntu-latest, macos-latest]
     runs-on: ${{ matrix.platform }}
     steps: