--- a/.github/workflows/golangci-lint.yml
+++ b/.github/workflows/golangci-lint.yml
@@ -7,7 +7,9 @@ on:
 
 jobs:
   golangci:
-    name: lint
+    strategy:
+      matrix:
+        go-version: [1.16]
     runs-on: ubuntu-latest
     steps:
       - name: Install Go
@@ -33,12 +35,5 @@ jobs:
         with:
           # Required: the version of golangci-lint is required and must be specified without patch version: we always use the latest patch version.
           version: v1.29
-          
-          # Optional: working directory, useful for monorepos
-          # working-directory: some_dir
-
           # Optional: golangci-lint command line arguments.
           args: --disable errcheck
-
-          # Optional: show only new issues if it's a pull request. The default value is `false`.
-          # only-new-issues: true