--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/relab/gorums
 
-go 1.15
+go 1.16
 
 require (
 	github.com/golang/protobuf v1.4.3