… benchmark/benchmark.pb.go
… benchmark/benchmark_gorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_async_gorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_correctable_gorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_multicast_gorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_qspec_gorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_quorumcall_gorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_rpc_gorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_server_gorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_types_gorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_unicast_gorums.pb.go
… examples/storage/proto/storage.pb.go
… examples/storage/proto/storage_gorums.pb.go
… go.sum
… gorums.pb.go
… internal/correctable/opts.pb.go
… ordering/ordering.pb.go