This PR replaces the deprecated golang.org/x/tools/go/loader with golang.org/x/tools/go/packages.
This had the nice benefit of a significant performance boost, which will be handy for productivity :-)

Before:
```sh
% time protoc-gen-gorums --bundle=./cmd/protoc-gen-gorums/gengorums/template_static.go
Generating bundle file: ./cmd/protoc-gen-gorums/gengorums/template_static.go
protoc-gen-gorums   81.15s user 157.17s system 1553% cpu 15.340 total
protoc-gen-gorums   82.20s user 157.40s system 1401% cpu 17.102 total
protoc-gen-gorums   80.74s user 154.08s system 1373% cpu 17.093 total
```

After:
```sh
% time protoc-gen-gorums --bundle=./cmd/protoc-gen-gorums/gengorums/template_static.go
Generating bundle file: ./cmd/protoc-gen-gorums/gengorums/template_static.go
protoc-gen-gorums   1.46s user 2.16s system 519% cpu 0.696 total
protoc-gen-gorums   1.49s user 2.31s system 528% cpu 0.719 total
protoc-gen-gorums   1.49s user 2.30s system 535% cpu 0.708 total
```

In addition, the PR simplifies and reorganizes the bundle functionality for clarity.

Fixes #130.